feat: rename xAI Grok OAuth to xAI Grok Build and update related definitions and documentation
