+++ src/client/definitions.ts
@@ -12,7 +12,7 @@ import { OllamaProvider } from './ollama/client';
 import { OpenAIChatCompletionProvider } from './openai/chat-completion-client';
 import { OpenAICodexProvider } from './openai/codex-client';
 import { OpenAIResponsesProvider } from './openai/responses-client';
-import { XaiGrokOAuthProvider } from './openai/xai-grok-oauth-client';
+import { XaiGrokBuildProvider } from './xai/grok-build-client';
 import { Feature } from './types';
 import { matchProvider, matchModelFamily } from './utils';
 
@@ -26,7 +26,7 @@ export type ProviderType =
   | 'github-copilot'
   | 'openai-chat-completion'
   | 'openai-codex'
-  | 'xai-grok-oauth'
+  | 'xai-grok-build'
   | 'openai-responses'
   | 'ollama';
 
@@ -108,12 +108,12 @@ export const PROVIDER_TYPES: Record<ProviderType, ProviderDefinition> = {
     category: 'Experimental',
     class: OpenAICodexProvider,
   },
-  'xai-grok-oauth': {
-    type: 'xai-grok-oauth',
-    label: t('xAI Grok OAuth (SuperGrok)'),
-    description: 'api.x.ai Responses (SuperGrok / X Premium+ OAuth bearer)',
+  'xai-grok-build': {
+    type: 'xai-grok-build',
+    label: t('xAI Grok Build'),
+    description: '/v1/responses',
     category: 'Experimental',
-    class: XaiGrokOAuthProvider,
+    class: XaiGrokBuildProvider,
   },
 };
